Move package, remove old file
