+++ drivers/robotpy_pybind_build_info.bzl
@@ -74,7 +74,7 @@ def drivers_extension(srcs = [], header_to_dat_deps = [], extra_hdrs = [], inclu
             class_name = "GoBildaPinpoint",
             yml_file = "semiwrap/GoBildaPinpoint.yml",
             header_root = "$(execpath :robotpy-native-wpilib-drivers.copy_headers)",
-            header_file = "$(execpath :robotpy-native-wpilib-drivers.copy_headers)/wpi/drivers/GoBildaPinpoint.hpp",
+            header_file = "$(execpath :robotpy-native-wpilib-drivers.copy_headers)/wpi/drivers/imu/GoBildaPinpoint.hpp",
             tmpl_class_names = [],
             trampolines = [
                 ("wpi::GoBildaPinpoint", "wpi__GoBildaPinpoint.hpp"),
