+++ drivers/src/main/java/org/wpilib/drivers/imu/package-info.java
@@ -1,0 +1,6 @@
+// Copyright (c) FIRST and other WPILib contributors.
+// Open Source Software; you can modify and/or share it under the terms of
+// the WPILib BSD license file in the root directory of this project.
+
+/** WPILib inertial measurement unit drivers. */
+package org.wpilib.drivers.imu;
