+++ drivers/src/main/native/cpp/imu/GoBildaPinpoint.cpp
@@ -6,7 +6,7 @@
 // This file is based on the goBILDA FTC driver, which was made available under
 // the MIT License. Copyright (c) 2025 Base 10 Assets, LLC.
 
-#include "wpi/drivers/GoBildaPinpoint.hpp"
+#include "wpi/drivers/imu/GoBildaPinpoint.hpp"
 
 #include <algorithm>
 #include <bit>
