+++ drivers/src/test/native/cpp/imu/GoBildaPinpointTest.cpp
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-#include "wpi/drivers/GoBildaPinpoint.hpp"
+#include "wpi/drivers/imu/GoBildaPinpoint.hpp"
 
 #include <algorithm>
 #include <bit>
